升级到25w03a

+++ src/main/java/org/carpetorgaddition/mixin/rule/HayBlockMixin.java
@@ -16,7 +16,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(HayBlock.class)
 public class HayBlockMixin {
     @Inject(method = "onLandedUpon", at = @At("HEAD"), cancellable = true)
-    private void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, float fallDistance, CallbackInfo ci) {
+    private void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, double fallDistance, CallbackInfo ci) {
         if (CarpetOrgAdditionSettings.hayBlockCompleteOffsetFall) {
             ci.cancel();
         }
